Add server-side validation to conditional fields …if a containing checkbox is selected (i.e. present in `form.cleaned_data`), contained subfields become required so that it's not only JS that can ensure they work correctly
TODO: perhaps these conditional fields need to remove their subfields' values from `self.cleaned_data` if not selected in case javascript is not enabled. The situation being: something is entered into a subfield but the containing checkbox is not checked. With javascript, the subfields become disabled so are not sent to the backend so are effectively treated as empty. Without javascript, the subfields remain enabled so backend will treat it as though the checkbox was selected.